divide hooks to avoid duplication of calculations and redux calls
