+++ src/components/ens-registration/RegistrationAvatar/RegistrationAvatar.tsx
@@ -17,7 +17,7 @@ import {
 import { UniqueAsset } from '@rainbow-me/entities';
 import { UploadImageReturnData } from '@rainbow-me/handlers/pinata';
 import {
-  useENSRegistration,
+  useENSAssignRegistration,
   useENSRegistrationForm,
   useSelectImageMenu,
 } from '@rainbow-me/hooks';
@@ -42,7 +42,7 @@ const RegistrationAvatar = ({
 }) => {
   const {
     images: { avatarUrl: initialAvatarUrl },
-  } = useENSRegistration();
+  } = useENSAssignRegistration();
   const {
     isLoading,
     values,
